Ranger Modernization & Tweaks (#3313)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Updates the Ranger-Class, giving it facilities that feel better without
significantly altering the balance.

![image](https://github.com/user-attachments/assets/b9a6de86-3fc3-44cb-9b9a-46c971fb5165)

Major changes: 
Adds a Logistics Specialist to the ship's crew manifest, and enhances
the Cargo Bay to accommodate for them. Now they have a desk, a bed, and
a locker, all painted in N+S colors! The Cargo Bay now also requires
ship access to be opened/closed.
The ship is now a bit more focused on "disaster response" rather than
being a generalist, and the tags have been updated to reflect this. I
might do some more work on this in the future, to lean into the "ship
that can fix both the crew and the ship" aspect. This also leaves the
niche of Vigilitas Gunship open for when that's made.
Re-paints the ship to be more in line with Vigilitas colors, and reflect
NT Spaceworks' designs (titanium hulls/windows) & updates the prefix to
VISV.

Smaller changes:
Adds access locks to the AI core/Comms (engineering access), the Ready
Room (requires Ship Access now and a non-intern role), security office,
and specialist quarters/rooms
Adds a few internal windows - the Bridge now has a shutter-able window
to the atrium, Security gets shutterable windows, and so does Medical.
Updates the Air Cycler to be more secure and not have any exterior
windows, keeping with the Ranger's "no external windows" theme.
Moves the atmospherics equipment to the engine bays, replacing the
equipment room with a proper ready-room for the LP Squad, and a small
armory compartment. Also adds a 9mm Saber SMG to the armory. All of the
LP hardsuits have been moved into this ready-room, except for the
Lieutenant's.
Enhances Medical by replacing the weird cell with a proper surgical
compartment, and uses tinted windows to give the Medical Specialist's
bedroom some privacy.
Rearranges the Security Specialist's office by giving them a proper
desk, a records console, and a camera monitor, aswell as a much more
functional holding cell.
Makes the AI core & Comms room much prettier (and more secure)
Adjusts the location of the docking port so you no longer need to cross
hyperspace
Expands the bridge slightly & adds more shutters to the ship.
Fixes the LP med-spec's hardsuit to be spaceproof again. 
Enhances the LT's living quarters.
Upgrades the Ion Engines to the premium variant - currently, it has the
same amount of thrusters as a Kilo, but is significantly heavier. As
it'd require significant hull changes to add more thrusters, they've
just been upgraded to make it more fitting for a proper response ship.
Adds an air refilling system, so you don't have to depressurize your air
supply in order to refill it.
Replaces the "Emergency EVA" closets with Fragile Space Suit Storage
Units, which are weaker than the previous EVA suits + spawns less
supplies.
Makes it so the LP Lieutenant spawns with the Vigilitas Security Coat
rather than the standard NT one, and removes the indie officer stuff
from their closet.
Fixes the Medical Specialist's hardsuit to be spaceproof again. 
Adds GPSes to every specialist's locker, instead of just a few, and
gives the AI Core a stationary computer GPS.
Shoves a dresser into Cryo.
Adds External Atmospherics control - you can now control the external
atmospherics fittings and also have an ability to actually check the
external atmosphere with sensors. Code for it's a bit jank but it works.
Are we out of scope yet?
Countless other small changes I've forgotten about as this expands.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes the Ranger feel much better to play on - brings the crew together
more, makes the areas feel a bit more usable, and helps the Ranger lean
into a specialization, rather than being Yet Another NT Generalist.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: Modernizes and slightly buffs the Ranger. 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: fighterslam <56803762+fighterslam@users.noreply.github.com>
Co-authored-by: Sadhorizon <108196626+Sadhorizon@users.noreply.github.com>